pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leap-data-management-utils"
dynamic = ["version"]
requires-python = ">=3.9"
description = "LEAP / pangeo-forge-recipes extension library for logging data in Google Big Query"
readme = "README.md"
authors = [{ name = "LEAP", email = "leap@columbia.edu" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "apache-beam",
    "cftime",
    "db_dtypes",
    "google-api-core",
    "google-cloud-bigquery",
    "pandas",
    "pangeo-forge-esgf",
    "pydantic-core",
    "pydantic>=2",
    "pyyaml",
    "universal-pathlib",
    "xarray",
    "zarr",
]

[project.optional-dependencies]
test = [
    "pre-commit",
]


[project.scripts]
leap-catalog = "leap_data_management_utils.catalog:main"


[tool.setuptools.packages.find]
include = ["leap_data_management_utils*"]

[tool.setuptools_scm]
write_to = "leap_data_management_utils/_version.py"
write_to_template = "__version__ = '{version}'"

[tool.ruff]
line-length = 100
target-version = "py39"
extend-include = ["*.ipynb"]


builtins = ["ellipsis"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
[tool.ruff.lint]
per-file-ignores = {}
ignore = [
    "E721", # Comparing types instead of isinstance
    "E741", # Ambiguous variable names
    "E501", # Conflicts with ruff format
]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I",
    # Pyupgrade
    "UP",
]



[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.ruff.lint.isort]
known-first-party = ["leap_data_management_utils"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"
